% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nguL_to_gL.R
\name{nguL_to_gL}
\alias{nguL_to_gL}
\title{To solve microliters needed for desired concentration of RNA/DNA}
\usage{
nguL_to_gL(ng_per_uL)
}
\arguments{
\item{ng_per_uL}{Input value, must be numerical value representing sample concentration.}
}
\value{
A numerical vector, with a sample's concentration in g/L.
}
\description{
This function takes a numerical input for the concentration
(ng/µL) of your RNA/DNA sample and returns the concentration in g/L.
}
