% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_pipette.R
\name{set_pipette_to}
\alias{set_pipette_to}
\title{This function is based off the simple equation: c1v1 = c2v2}
\usage{
set_pipette_to(c1, v1, c2)
}
\arguments{
\item{c1}{Input value, must be numerical value representing initial concentration.}

\item{v1}{Input value, must be numerical value representing initial volume.}

\item{c2}{Input calue, must be numerical value representing final concentration.}
}
\value{
A numerical vector, with the volume to set your micro-pipette.
}
\description{
This function requires three inputs: c1, v1, c2.
    c1 = initial concentration of sample (g/L; output of first function)
    v2 = initial volume of the sample (g/L)
    c2 = the desired concentration of the sample (g/L)
}
\details{
This function solves for v2 (the final volume of the desired concentration, c2)
then it finds the absolute difference between v1 and v2
(the volume of liquid needed in the sample to reach c2). The function the converts
that "volume to add" in to ng/µL, which will be the numerical value to set your pipette
to reach the final concentration, c2
}
